requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
email-validator
pymysql
python-dotenv
bcrypt==3.2.0
torch
transformers
scipy
pandas
scikit-learn
torchvision
pillow
fastapi-mail
google-auth
requests
